Verificacion de Jerarquia casi completada en el tema Usuarios
Falta no dejar que el empleado se suba a administrador xd

+++ PaginaGrupo2/modelo/eliminar_Usuario_modelo.php
@@ -9,9 +9,24 @@ class Eliminar_usuario{
     }
 
     public function eliminar_usuario ($id) {
-    $sql = "DELETE FROM usuario WHERE id_usuario = $id";
-    $query = mysqli_query($this->con, $sql);
-    return $query;
+    session_start(); 
+    $permiso = $_SESSION['tipo'];
+    $id_personal = $_SESSION['id'];
+    $verificacion = "SELECT * FROM usuario WHERE id_usuario = '$id'";
+    $query_veri = mysqli_query($this->con, $verificacion);
+    $tipo = mysqli_fetch_array($query_veri);
+    if($tipo['tipo'] != "administrador" && $permiso == "empleado"){
+        if($id_personal != $id){
+        $sql = "DELETE FROM usuario WHERE id_usuario = $id";
+        $query = mysqli_query($this->con, $sql);
+        return $query;
+    }else{
+        $estado=3;
+        return $estado;
+    }
+    }else{
+        return;
+    }
     }
 
     }
